refactor: 서비스 레이어를 types와 api 레이어로 분리
